Human Benchmark Bots/TypeX/requirements.txt
mss
pillow
pyautogui
keyboard
pytesseract
time
re